ART: Optimized implementation of System.arraycopy(char) intrinsic
This change introduces loop inversion and loop unrolling
optimizations to the System.arraycopy(char) intrinsic.
Instead of a single loop with a pair of strh/ldrh instructions
we now have two loops: one with ldr/str instructions and
the other one with strh/ldrh. The goal is to use increased
step size (8 bytes) in the first loop and then after
it's over finish processing of the rest of the array
using the smaller step. We also increased the threshold
array size when intrinsic implementation is chosen.
Also added tests to cover arraycopy(char) case for
arrays of different sizes.

Test: 011-array-copy
Test: test-art-target
Change-Id: I45b196789bdb0906bb1c79ae0784d347cc75aee4
Signed-off-by: Jason Edson <jaysonedson@gmail.com>